pyproject.toml
[tool.poetry]
name = "ai-hub-discord-bot"
version = "1.0.0"
description = "ChatGPT、BingGPT、Image Creator、DALL_E、Other AI Application with Discord Bot"
authors = ["Lin-Rexter <aaaa35963596@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_hub_discord_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
revchatgpt = "^4.0.9"
edgegpt = "^0.1.10.1"
openai = "^0.27.2"
requests = "^2.28.2"
deep-translator = "^1.10.1"
py-cord = "^2.4.1"
numpy = "^1.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
